+++ src/app.html
@@ -11,16 +11,14 @@
 	</body>
 	<style>
 		:root {
-			/* --behind-color: rgb(227, 236, 243); */
-			--behind-color: rgb(255, 255, 255);
+			--behind-color: rgb(227, 236, 243);
 			--background-color: rgb(255, 255, 255);
 			--accent-color: rgb(31, 145, 238);
 			--font-color: rgb(51, 45, 45);
 		}
 		@media (prefers-color-scheme: dark) {
 			:root {
-				/* --behind-color: rgb(16, 20, 27); */
-				--behind-color: rgb(35, 34, 34);
+				--behind-color: rgb(16, 20, 27);
 				--background-color: rgb(35, 34, 34);
 				--accent-color: rgb(111, 79, 252);
 				--font-color: rgb(192, 179, 179);
@@ -50,7 +48,7 @@
 			border-color: var(--accent-color);
 		}
 		h1, h2, h3, h4, h5, h6 {
-			text-decoration: underline var(--behind-color);
+			text-decoration: underline wavy var(--behind-color);
 		}
 	</style>
 </html>
