+++ src/routes/restricted/conseils/+page.svelte
@@ -1,8 +1,7 @@
 
 
 <script>
-    import Banner from "$lib/components/Banner.svelte";
-import Markdown from "$lib/components/Markdown.svelte";
+    import Markdown from "$lib/components/Markdown.svelte";
     import Rolldown from "$lib/components/Rolldown.svelte";
 
 
@@ -10,13 +9,10 @@ export let data;
 
 </script>
 
-<Banner/>
 
 <h1>Conseils des anciens</h1>
 
 <p>Conseils des anciens du parcours listés par l'ordre de plus récent (année fin L3).</p>
-
-<br/>
 
 {#await data.streamed.descs}
 Chargement...
